+++ tests/Tests.py
@@ -1,5 +1,5 @@
 import unittest
-from src.Solver import Solver
+from prologpy.Solver import Solver
 
 # Integration tests which test some of the higher level functionality of our implementation
 class PrologTestCases(unittest.TestCase):
